Refactors getting the month index and adds a shared getMonthData function
