+++ src/pages/DayDisplay/DayDisplay.tsx
@@ -1,6 +1,7 @@
 import { Link } from "react-router-dom";
 import Day from "../../components/Day/Day";
 import { formatDate } from "../../utils/date-utils";
+import { getMonthData } from "../../utils/month-utils";
 
 const DayDisplay = () => {
   const today = new Date();
@@ -8,22 +9,7 @@ const DayDisplay = () => {
 
   let index = today.getMonth();
 
-  const months = [
-    { name: "January", numberOfDays: 31 },
-    { name: "February", numberOfDays: 28 },
-    { name: "March", numberOfDays: 31 },
-    { name: "April", numberOfDays: 30 },
-    { name: "May", numberOfDays: 31 },
-    { name: "June", numberOfDays: 30 },
-    { name: "July", numberOfDays: 31 },
-    { name: "August", numberOfDays: 31 },
-    { name: "September", numberOfDays: 30 },
-    { name: "October", numberOfDays: 31 },
-    { name: "November", numberOfDays: 30 },
-    { name: "December", numberOfDays: 31 },
-  ];
-
-  const month = months[index];
+  const month = getMonthData(index);
 
   return (
     <main>
