+++ src/pages/MonthDisplay/MonthDisplay.tsx
@@ -1,5 +1,6 @@
 import { useState } from "react";
 import { useParams } from "react-router-dom";
+import { getMonthData } from "../../utils/month-utils";
 import Month from "../../components/Month/Month";
 import Modal from "../../components/Modal/Modal";
 
@@ -7,26 +8,13 @@ const MonthDisplay = () => {
   const today = new Date();
   const year = today.getFullYear();
 
-  const defaultIndex = Number(useParams().index);
-  let index = defaultIndex ? defaultIndex : today.getMonth();
+  const defaultIndex = useParams().index;
+  let index = Number(defaultIndex ?? today.getMonth());
 
   const [monthIndex, setMonthIndex] = useState(index);
   const [modalDate, setModalDate] = useState("");
 
-  const months = [
-    { name: "January", numberOfDays: 31 },
-    { name: "February", numberOfDays: 28 },
-    { name: "March", numberOfDays: 31 },
-    { name: "April", numberOfDays: 30 },
-    { name: "May", numberOfDays: 31 },
-    { name: "June", numberOfDays: 30 },
-    { name: "July", numberOfDays: 31 },
-    { name: "August", numberOfDays: 31 },
-    { name: "September", numberOfDays: 30 },
-    { name: "October", numberOfDays: 31 },
-    { name: "November", numberOfDays: 30 },
-    { name: "December", numberOfDays: 31 },
-  ];
+  const month = getMonthData(monthIndex);
 
   const updateMonthIndex = (value: number): void => {
     const newMonthIndex = (monthIndex + value) % 12;
@@ -44,9 +32,9 @@ const MonthDisplay = () => {
   return (
     <main>
       <Month
-        name={months[monthIndex].name}
+        name={month.name}
         year={year}
-        numberOfDays={months[monthIndex].numberOfDays}
+        numberOfDays={month.numberOfDays}
         updateMonthIndex={updateMonthIndex}
         openModal={openModal}
       />
